remove interceptor for now, might cause issues later. update all routes to use auth

+++ src/tasks/tasks.service.ts
@@ -14,12 +14,14 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) {
       throw new NotFoundException();
     }
@@ -30,29 +32,33 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number) {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User) {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (!result.affected) {
       throw new NotFoundException();
     }
     return;
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async deleteAllTasks(filterDto: GetTasksFilterDto): Promise<void> {
-    const tasks = await this.taskRepository.getTasks(filterDto);
-    for (const task of tasks) {
-      const result = await this.taskRepository.delete(task.id);
-      if (!result.affected) {
-        throw new NotFoundException();
-      }
-    }
-    return;
-  }
+  // async deleteAllTasks(filterDto: GetTasksFilterDto): Promise<void> {
+  //   const tasks = await this.taskRepository.getTasks(filterDto);
+  //   for (const task of tasks) {
+  //     const result = await this.taskRepository.delete(task.id);
+  //     if (!result.affected) {
+  //       throw new NotFoundException();
+  //     }
+  //   }
+  //   return;
+  // }
 }
